refactor: centralize nested package path resolution, fix release facts
Extract shared nested-path mapping into scripts/lib/resolve-package-path.ts.
Both check-publish-closure.ts and check-manifest-invariants.ts now consume
the single resolver instead of maintaining duplicate mapping tables.

Update facts.json: published_packages 35 -> 36, build_targets 97 -> 98
(new adapter-managed-agents package).

Remove pending-registration-v0129.json from this branch (belongs to
conformance PR, not adapter PR).
