+++ scripts/check-publish-closure.ts
@@ -12,6 +12,7 @@
 
 import * as fs from 'fs';
 import * as path from 'path';
+import { resolvePackagePath } from './lib/resolve-package-path.js';
 
 const ROOT = path.resolve(__dirname, '..');
 const MANIFEST_PATH = path.join(ROOT, 'scripts/publish-manifest.json');
@@ -22,64 +23,10 @@ interface PublishManifest {
 
 /**
  * Map npm package name to directory path.
- * Handles both flat (packages/kernel) and nested (packages/adapters/openclaw) layouts.
+ * Uses shared resolver from scripts/lib/resolve-package-path.ts.
  */
 function packageDir(npmName: string): string {
-  // Strip @peac/ prefix
-  const shortName = npmName.replace('@peac/', '');
-
-  // Check flat layout first: packages/<shortName>
-  const flatDir = path.join(ROOT, 'packages', shortName);
-  if (fs.existsSync(path.join(flatDir, 'package.json'))) {
-    return flatDir;
-  }
-
-  // Check nested layouts.
-  // INVARIANT: Every new nested package (packages/adapters/*, packages/mappings/*,
-  // packages/rails/*, packages/capture/*, packages/transport/*, packages/net/*)
-  // must be added here. The flat-path fallback only works for top-level packages
-  // like packages/kernel. Failure to add an entry causes the publish-manifest
-  // closure check to fail in CI with "package.json not found."
-  const nestedMappings: Record<string, string> = {
-    'adapter-core': 'packages/adapters/core',
-    'adapter-openclaw': 'packages/adapters/openclaw',
-    'adapter-x402': 'packages/adapters/x402',
-    'adapter-x402-daydreams': 'packages/adapters/x402/daydreams',
-    'adapter-x402-fluora': 'packages/adapters/x402/fluora',
-    'adapter-x402-pinata': 'packages/adapters/x402/pinata',
-    'adapter-managed-agents': 'packages/adapters/managed-agents',
-    'adapter-openai-compatible': 'packages/adapters/openai-compatible',
-    'adapter-did': 'packages/adapters/did',
-    'adapter-eat': 'packages/adapters/eat',
-    'rails-x402': 'packages/rails/x402',
-    'rails-stripe': 'packages/rails/stripe',
-    'rails-card': 'packages/rails/card',
-    'rails-razorpay': 'packages/rails/razorpay',
-    'mappings-mcp': 'packages/mappings/mcp',
-    'mappings-a2a': 'packages/mappings/a2a',
-    'mappings-acp': 'packages/mappings/acp',
-    'mappings-aipref': 'packages/mappings/aipref',
-    'mappings-rsl': 'packages/mappings/rsl',
-    'mappings-tap': 'packages/mappings/tap',
-    'mappings-ucp': 'packages/mappings/ucp',
-    'mappings-content-signals': 'packages/mappings/content-signals',
-    'mappings-paymentauth': 'packages/mappings/paymentauth',
-    'capture-core': 'packages/capture/core',
-    'capture-node': 'packages/capture/node',
-    'sdk-js': 'packages/sdk-js',
-    'net-node': 'packages/net/node',
-    'transport-grpc': 'packages/transport/grpc',
-    'mappings-intoto': 'packages/mappings/intoto',
-    'mappings-slsa': 'packages/mappings/slsa',
-    disc: 'packages/discovery',
-  };
-
-  if (nestedMappings[shortName]) {
-    return path.join(ROOT, nestedMappings[shortName]);
-  }
-
-  // Fallback: try the flat name
-  return flatDir;
+  return resolvePackagePath(npmName);
 }
 
 function main() {
