--- specs/conformance/pending-registration-v0129.json
@@ -1,63 +1,0 @@
-{
-  "description": "Authoritative reconciliation of pending conformance ID registration for v0.12.9 PR 6. Source: check-matrix.mjs PENDING_REGISTRATION set, verified by extraction Apr 10, 2026.",
-  "current_registered": 192,
-  "pending_count": 25,
-  "target_post_pr6": 217,
-  "namespaces": {
-    "DID-RES": {
-      "count": 7,
-      "ids": [
-        "DID-RES-001",
-        "DID-RES-002",
-        "DID-RES-003",
-        "DID-RES-004",
-        "DID-RES-005",
-        "DID-RES-006",
-        "DID-RES-007"
-      ],
-      "source": "v0.12.6 adapter-did",
-      "governing_spec": "docs/specs/DID-PROFILE.md"
-    },
-    "GRPC-META": {
-      "count": 3,
-      "ids": ["GRPC-META-001", "GRPC-META-002", "GRPC-META-003"],
-      "source": "v0.12.6 transport-grpc",
-      "governing_spec": "docs/specs/GRPC-CARRIER.md"
-    },
-    "PKCE": {
-      "count": 4,
-      "ids": ["PKCE-001", "PKCE-002", "PKCE-003", "PKCE-004"],
-      "source": "v0.12.6 A2A OAuth",
-      "governing_spec": "docs/specs/MCP-AUTH-PROFILE.md"
-    },
-    "RURL": {
-      "count": 4,
-      "ids": ["RURL-001", "RURL-002", "RURL-003", "RURL-004"],
-      "source": "v0.12.6 receipt-url middleware",
-      "governing_spec": "docs/specs/EVIDENCE-CARRIER-CONTRACT.md"
-    },
-    "SC": {
-      "count": 3,
-      "ids": ["SC-001", "SC-003", "SC-004"],
-      "note": "SC-002 absent from pending set",
-      "source": "v0.12.6 mappings-intoto, mappings-slsa",
-      "governing_spec": "docs/specs/SUPPLY-CHAIN-PROFILE.md"
-    },
-    "X402V2": {
-      "count": 4,
-      "ids": ["X402V2-001", "X402V2-002", "X402V2-003", "X402V2-004"],
-      "source": "v0.12.6 adapter-x402 V2",
-      "governing_spec": "docs/specs/X402-V2-PROFILE.md"
-    }
-  },
-  "implementation_sequence": [
-    "1. Add 25 requirements to REQUIREMENTS array in scripts/conformance/build-registry.mjs",
-    "2. Regenerate requirement-ids.json via node scripts/conformance/build-registry.mjs",
-    "3. Add source-fragment annotations to governing spec docs",
-    "4. Update fixture inventory and test mappings",
-    "5. Regenerate matrix via node scripts/conformance/generate-matrix.mjs",
-    "6. REMOVE all 25 entries from PENDING_REGISTRATION in check-matrix.mjs",
-    "7. Verify: node scripts/conformance/check-matrix.mjs exits 0 with 0 exemptions",
-    "8. Verify: node scripts/conformance/verify-registry-drift.mjs exits 0"
-  ]
-}
